Working Project with AndroidX updated dependency

+++ imageeditengine/src/main/java/com/droidninja/imageeditengine/BaseImageEditActivity.java
@@ -2,21 +2,22 @@ package com.droidninja.imageeditengine;
 
 import android.graphics.Matrix;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
 public abstract class BaseImageEditActivity extends AppCompatActivity {
 
-  @Override protected void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-  }
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
 
-  public static int[] getBitmapOffset(ImageView img,  Boolean includeLayout) {
-    int[] offset = new int[2];
-    float[] values = new float[9];
+    public static int[] getBitmapOffset(ImageView img, Boolean includeLayout) {
+        int[] offset = new int[2];
+        float[] values = new float[9];
 
     Matrix m = img.getImageMatrix();
     m.getValues(values);
@@ -25,12 +26,12 @@ public abstract class BaseImageEditActivity extends AppCompatActivity {
     offset[1] = (int) values[2];
 
     if (includeLayout) {
-      ViewGroup.MarginLayoutParams lp = (ViewGroup.MarginLayoutParams) img.getLayoutParams();
-      int paddingTop = (int) (img.getPaddingTop() );
-      int paddingLeft = (int) (img.getPaddingLeft() );
+        ViewGroup.MarginLayoutParams lp = (ViewGroup.MarginLayoutParams) img.getLayoutParams();
+        int paddingTop = img.getPaddingTop();
+        int paddingLeft = img.getPaddingLeft();
 
-      offset[0] += paddingTop + lp.topMargin;
-      offset[1] += paddingLeft + lp.leftMargin;
+        offset[0] += paddingTop + lp.topMargin;
+        offset[1] += paddingLeft + lp.leftMargin;
     }
     return offset;
   }
